Pequenos ajustes no código

+++ entidade/Cliente.java
@@ -3,7 +3,7 @@ package entidade;
 public class Cliente {
 
 	public String nome;
-	public long telefone;
+	public int telefone;
 	public int rg;
